Moved Category enum to category package
