+++ backend/src/main/java/com/java55/supermarktitvitae/category/Category.java
@@ -1,4 +1,4 @@
-package com.java55.supermarktitvitae.product;
+package com.java55.supermarktitvitae.category;
 
 public enum Category {
     GROENTE,
